requirements.txt
discord.py
pynacl
python-dotenv
yt-dlp
ffmpeg
asyncio
